data_preparation/requirements.txt
pillow
opencv-python
numpy
flask
flask_restful
flask_cors
flask-swagger-ui
